# Optimize Affine3D using 4x Vec3A