This makes the SSE vector transforms faster, which argueably is the most important. But a lot of other numbers gets slightly worse (including memory usage).


| op                    | old sse2 | new sse2 |   | old scalar | new scalar |
| --------------------- | -------- | -------- | - | ---------- | ---------- |
| `inverse`             | `9.3 ns` | `8.4 ns` 🥇 |   |  `16 ns`   | `16 ns`   |
| `Affine3D * Affine3D` | `4.2 ns` 🥇 | `5.2 ns` |   |  `13 ns`   | `11 ns` 🥇   |
| `transform point3`    | `3.8 ns` | `2.7 ns` 🥇 |   |  `3.5 ns` 🥇  | `4.0 ns`   |
| `transform vector3`   | `3.6 ns` | `2.5 ns` 🥇 |   |  `2.9 ns` 🥇  | `3.6 ns`   |
| size                  | 48 bytes 🥇 | 64 bytes |   | 48 bytes 🥇 | 64 bytes |