@bitshifter 2021-04-12
I guess the scalar multiply is losing out due to the alignment/padding :(